+++ shuup/campaigns/utils/campaigns.py
@@ -14,4 +14,6 @@ def get_product_ids_and_quantities(basket):
         if not line.product:
             continue
         q_counter[line.product.id] += line.quantity
+        if line.product.variation_parent_id:
+            q_counter[line.product.variation_parent_id] += line.quantity
     return dict(q_counter.most_common())