+++ shuup_tests/campaigns/test_basket_product_conditions.py
@@ -12,6 +12,7 @@ from shuup.campaigns.models.basket_conditions import (
     BasketTotalAmountCondition, BasketTotalProductAmountCondition,
     ComparisonOperator, ProductsInBasketCondition
 )
+from shuup.core.models import ProductMode
 from shuup.front.basket import get_basket
 from shuup.testing.factories import create_product, get_default_supplier
 from shuup_tests.campaigns import initialize_test
@@ -49,6 +50,36 @@ def test_product_in_basket_condition(rf):
 
     basket.add_product(supplier=supplier, shop=shop, product=product, quantity=1)
     assert not condition.matches(basket, [])
+
+
+@pytest.mark.django_db
+def test_product_in_basket_condition_with_variation_parent(rf):
+    request, shop, group = initialize_test(rf, False)
+
+    basket = get_basket(request)
+    supplier = get_default_supplier()
+
+    product = create_product(
+        "test-product", shop, default_price="200", supplier=supplier, mode=ProductMode.SIMPLE_VARIATION_PARENT)
+
+    child_products = []
+    for x in range(0, 3):
+        child_product = create_product("test-product-%s" % x, shop, default_price="10", supplier=supplier)
+        child_product.link_to_parent(product)
+        child_products.append(child_product)
+
+    condition = ProductsInBasketCondition.objects.create()
+    condition.values = [product]
+    condition.operator = ComparisonOperator.EQUALS
+    condition.quantity = 3
+    condition.save()
+
+    assert not condition.matches(basket, [])
+
+    for child_product in child_products:
+        basket.add_product(supplier=supplier, shop=shop, product=child_product, quantity=1)
+
+    assert condition.matches(basket, [])
 
 
 @pytest.mark.django_db